+++ src/fastfield/multivalued/mod.rs
@@ -6,8 +6,6 @@ pub use self::writer::MultiValuedFastFieldWriter;
 
 #[cfg(test)]
 mod tests {
-
-    use chrono::Duration;
     use proptest::strategy::Strategy;
     use proptest::{prop_oneof, proptest};
     use test_log::test;
@@ -16,7 +14,9 @@ mod tests {
     use crate::indexer::NoMergePolicy;
     use crate::query::QueryParser;
     use crate::schema::{Cardinality, Facet, FacetOptions, NumericOptions, Schema};
-    use crate::{Document, Index, Term};
+    use crate::time::format_description::well_known::Rfc3339;
+    use crate::time::{Duration, OffsetDateTime};
+    use crate::{DateTime, Document, Index, Term};
 
     #[test]
     fn test_multivalued_u64() -> crate::Result<()> {
@@ -69,22 +69,27 @@ mod tests {
         let schema = schema_builder.build();
         let index = Index::create_in_ram(schema);
         let mut index_writer = index.writer_for_tests()?;
-        let first_time_stamp = chrono::Utc::now();
-        index_writer.add_document(
-            doc!(date_field=>first_time_stamp, date_field=>first_time_stamp, time_i=>1i64),
-        )?;
-        index_writer.add_document(doc!(time_i=>0i64))?;
+        let first_time_stamp = OffsetDateTime::now_utc();
+        index_writer.add_document(doc!(
+                date_field => DateTime::new_utc(first_time_stamp),
+                date_field => DateTime::new_utc(first_time_stamp),
+                time_i=>1i64))?;
+        index_writer.add_document(doc!(time_i => 0i64))?;
         // add one second
-        index_writer.add_document(
-            doc!(date_field=>first_time_stamp + Duration::seconds(1), time_i=>2i64),
-        )?;
+        index_writer.add_document(doc!(
+            date_field => DateTime::new_utc(first_time_stamp + Duration::seconds(1)),
+            time_i => 2i64))?;
         // add another second
         let two_secs_ahead = first_time_stamp + Duration::seconds(2);
-        index_writer.add_document(doc!(date_field=>two_secs_ahead, date_field=>two_secs_ahead,date_field=>two_secs_ahead, time_i=>3i64))?;
+        index_writer.add_document(doc!(
+            date_field => DateTime::new_utc(two_secs_ahead),
+            date_field => DateTime::new_utc(two_secs_ahead),
+            date_field => DateTime::new_utc(two_secs_ahead),
+            time_i => 3i64))?;
         // add three seconds
-        index_writer.add_document(
-            doc!(date_field=>first_time_stamp + Duration::seconds(3), time_i=>4i64),
-        )?;
+        index_writer.add_document(doc!(
+                date_field => DateTime::new_utc(first_time_stamp + Duration::seconds(3)),
+                time_i => 4i64))?;
         index_writer.commit()?;
 
         let reader = index.reader()?;
@@ -96,7 +101,7 @@ mod tests {
             let parser = QueryParser::for_index(&index, vec![]);
             let query = parser.parse_query(&format!(
                 "multi_date_field:\"{}\"",
-                first_time_stamp.to_rfc3339()
+                first_time_stamp.format(&Rfc3339)?,
             ))?;
             let results = searcher.search(&query, &TopDocs::with_limit(5))?;
             assert_eq!(results.len(), 1);
@@ -107,9 +112,8 @@ mod tests {
                         .get_first(date_field)
                         .expect("cannot find value")
                         .as_date()
-                        .unwrap()
-                        .timestamp(),
-                    first_time_stamp.timestamp()
+                        .unwrap(),
+                    DateTime::new_utc(first_time_stamp),
                 );
                 assert_eq!(
                     retrieved_doc
@@ -123,7 +127,7 @@ mod tests {
 
         {
             let parser = QueryParser::for_index(&index, vec![date_field]);
-            let query = parser.parse_query(&format!("\"{}\"", two_secs_ahead.to_rfc3339()))?;
+            let query = parser.parse_query(&format!("\"{}\"", two_secs_ahead.format(&Rfc3339)?))?;
             let results = searcher.search(&query, &TopDocs::with_limit(5))?;
 
             assert_eq!(results.len(), 1);
@@ -135,9 +139,8 @@ mod tests {
                         .get_first(date_field)
                         .expect("cannot find value")
                         .as_date()
-                        .unwrap()
-                        .timestamp(),
-                    two_secs_ahead.timestamp()
+                        .unwrap(),
+                    DateTime::new_utc(two_secs_ahead)
                 );
                 assert_eq!(
                     retrieved_doc
@@ -153,8 +156,8 @@ mod tests {
             let parser = QueryParser::for_index(&index, vec![date_field]);
             let range_q = format!(
                 "multi_date_field:[{} TO {}}}",
-                (first_time_stamp + Duration::seconds(1)).to_rfc3339(),
-                (first_time_stamp + Duration::seconds(3)).to_rfc3339()
+                (first_time_stamp + Duration::seconds(1)).format(&Rfc3339)?,
+                (first_time_stamp + Duration::seconds(3)).format(&Rfc3339)?
             );
             let query = parser.parse_query(&range_q)?;
             let results = searcher.search(&query, &TopDocs::with_limit(5))?;
@@ -177,9 +180,8 @@ mod tests {
                         .get_first(date_field)
                         .expect("cannot find value")
                         .as_date()
-                        .expect("value not of Date type")
-                        .timestamp(),
-                    (first_time_stamp + Duration::seconds(offset_sec)).timestamp()
+                        .expect("value not of Date type"),
+                    DateTime::new_utc(first_time_stamp + Duration::seconds(offset_sec)),
                 );
                 assert_eq!(
                     retrieved_doc