+++ src/query/query_parser/query_parser.rs
@@ -15,8 +15,10 @@ use crate::query::{
 use crate::schema::{
     Facet, FacetParseError, Field, FieldType, IndexRecordOption, Schema, Term, Type,
 };
+use crate::time::format_description::well_known::Rfc3339;
+use crate::time::{OffsetDateTime, UtcOffset};
 use crate::tokenizer::{TextAnalyzer, TokenizerManager};
-use crate::Score;
+use crate::{DateTime, Score};
 
 /// Possible error that may happen when parsing a query.
 #[derive(Debug, PartialEq, Eq, Error)]
@@ -72,7 +74,7 @@ pub enum QueryParserError {
     RangeMustNotHavePhrase,
     /// The format for the date field is not RFC 3339 compliant.
     #[error("The date field has an invalid format")]
-    DateFormatError(#[from] chrono::ParseError),
+    DateFormatError(#[from] time::error::Parse),
     /// The format for the facet field is invalid.
     #[error("The facet field is malformed: {0}")]
     FacetFormatError(#[from] FacetParseError),
@@ -331,11 +333,8 @@ impl QueryParser {
                 Ok(Term::from_field_f64(field, val))
             }
             FieldType::Date(_) => {
-                let dt = chrono::DateTime::parse_from_rfc3339(phrase)?;
-                Ok(Term::from_field_date(
-                    field,
-                    &dt.with_timezone(&chrono::Utc),
-                ))
+                let dt = OffsetDateTime::parse(phrase, &Rfc3339)?;
+                Ok(Term::from_field_date(field, DateTime::new_utc(dt)))
             }
             FieldType::Str(ref str_options) => {
                 let option = str_options.get_indexing_options().ok_or_else(|| {
@@ -408,8 +407,8 @@ impl QueryParser {
                 Ok(vec![LogicalLiteral::Term(f64_term)])
             }
             FieldType::Date(_) => {
-                let dt = chrono::DateTime::parse_from_rfc3339(phrase)?;
-                let dt_term = Term::from_field_date(field, &dt.with_timezone(&chrono::Utc));
+                let dt = OffsetDateTime::parse(phrase, &Rfc3339)?;
+                let dt_term = Term::from_field_date(field, DateTime::new_utc(dt));
                 Ok(vec![LogicalLiteral::Term(dt_term)])
             }
             FieldType::Str(ref str_options) => {
@@ -665,12 +664,12 @@ enum NumValue {
     U64(u64),
     I64(i64),
     F64(f64),
-    DateTime(crate::DateTime),
+    DateTime(OffsetDateTime),
 }
 
 fn infer_type_num(phrase: &str) -> Option<NumValue> {
-    if let Ok(dt) = chrono::DateTime::parse_from_rfc3339(phrase) {
-        let dt_utc = dt.with_timezone(&chrono::Utc);
+    if let Ok(dt) = OffsetDateTime::parse(phrase, &Rfc3339) {
+        let dt_utc = dt.to_offset(UtcOffset::UTC);
         return Some(NumValue::DateTime(dt_utc));
     }
     if let Ok(u64_val) = str::parse::<u64>(phrase) {
@@ -712,7 +711,7 @@ fn generate_literals_for_json_object(
                 json_term_writer.set_fast_value(f64_val);
             }
             NumValue::DateTime(dt_val) => {
-                json_term_writer.set_fast_value(dt_val);
+                json_term_writer.set_fast_value(DateTime::new_utc(dt_val));
             }
         }
         logical_literals.push(LogicalLiteral::Term(json_term_writer.term().clone()));
@@ -1039,6 +1038,7 @@ mod test {
 
     #[test]
     fn test_json_field_possibly_a_date() {
+        // Subseconds are discarded
         test_parse_query_to_logical_ast_helper(
             r#"json.date:"2019-10-12T07:20:50.52Z""#,
             r#"(Term(type=Json, field=14, path=date, vtype=Date, 2019-10-12T07:20:50Z) "[(0, Term(type=Json, field=14, path=date, vtype=Str, "2019")), (1, Term(type=Json, field=14, path=date, vtype=Str, "10")), (2, Term(type=Json, field=14, path=date, vtype=Str, "12t07")), (3, Term(type=Json, field=14, path=date, vtype=Str, "20")), (4, Term(type=Json, field=14, path=date, vtype=Str, "50")), (5, Term(type=Json, field=14, path=date, vtype=Str, "52z"))]")"#,