+++ src/schema/term.rs
@@ -70,8 +70,8 @@ impl Term {
     }
 
     /// Builds a term given a field, and a DateTime value
-    pub fn from_field_date(field: Field, val: &DateTime) -> Term {
-        Term::from_fast_value(field, val)
+    pub fn from_field_date(field: Field, val: DateTime) -> Term {
+        Term::from_fast_value(field, &val)
     }
 
     /// Creates a `Term` given a facet.
@@ -126,7 +126,7 @@ impl Term {
     }
 
     /// Sets a `i64` value in the term.
-    pub fn set_date(&mut self, date: crate::DateTime) {
+    pub fn set_date(&mut self, date: DateTime) {
         self.set_fast_value(date);
     }
 
@@ -266,8 +266,8 @@ where B: AsRef<[u8]>
     ///
     /// Returns None if the term is not of the Date type, or if the term byte representation
     /// is invalid.
-    pub fn as_date(&self) -> Option<crate::DateTime> {
-        self.get_fast_type::<crate::DateTime>()
+    pub fn as_date(&self) -> Option<DateTime> {
+        self.get_fast_type::<DateTime>()
     }
 
     /// Returns the text associated with the term.
@@ -374,7 +374,7 @@ fn debug_value_bytes(typ: Type, bytes: &[u8], f: &mut fmt::Formatter) -> fmt::Re
         }
         // TODO pretty print these types too.
         Type::Date => {
-            write_opt(f, get_fast_type::<crate::DateTime>(bytes))?;
+            write_opt(f, get_fast_type::<DateTime>(bytes))?;
         }
         Type::Facet => {
             let facet_str = str::from_utf8(bytes)